@@ -6,6 +6,9 @@
   "author": "MoscowJS Team <team@moscowjs.org>",
   "dependencies": {
     "@directus/gatsby-source-directus": "^9.16.2",
+    "@types/node": "^22.5.0",
+    "@types/react": "^18.3.4",
+    "@types/react-dom": "^18.3.0",
     "@types/react-helmet": "^6.1.1",
     "@types/remarkable": "^2.0.2",
     "@types/styled-components": "^5.1.10",
@@ -45,7 +48,7 @@
     "remark-toc": "^7.2.0",
     "styled-components": "^5.3.0",
     "ts-node": "^10.0.0",
-    "typescript": "^4.3.4",
+    "typescript": "^4.9.5",
     "typography": "^0.16.21"
   },
   "devDependencies": {
