@@ -3,21 +3,30 @@
 //   speakerPath,
 // } = require("./src/utils/paths.ts")
 
-exports.onCreateWebpackConfig = ({ actions, stage, plugins }) => {
-  if (stage === "build-javascript" || stage === "develop") {
+import type { GatsbyNode } from 'gatsby'
+
+export const onCreateWebpackConfig: GatsbyNode['onCreateWebpackConfig'] = ({
+  actions,
+  stage,
+  plugins,
+}) => {
+  if (stage === 'build-javascript' || stage === 'develop') {
     actions.setWebpackConfig({
-      plugins: [plugins.provide({ process: "process/browser" })],
+      plugins: [plugins.provide({ process: 'process/browser' })],
     })
   }
 }
 
-exports.createPages = async ({ graphql, actions }) => {
+export const createPages: GatsbyNode['createPages'] = async ({
+  graphql,
+  actions,
+}) => {
   const { createPage } = actions
 
   const result = await graphql(`
     query OtherPlaylistInfoQuery {
       directus {
-        persons(limit:3) {
+        persons(limit: 3) {
           id
           name
           email
@@ -29,7 +38,6 @@ exports.createPages = async ({ graphql, actions }) => {
 
   console.log('----result', JSON.stringify(result, null, 2))
 
-
   // result.data.directus.persons.forEach((person) => {
   //   createPage({
   //     path: speakerPath(person.name),
@@ -37,5 +45,4 @@ exports.createPages = async ({ graphql, actions }) => {
   //     context: { id: person.id },
   //   })
   // })
-
 }
